Added invoice pdf print. Cannot delete contractor that is used on invoice. Pathfiles are now absolut. Still need refactor, mess with std:string and Qstring, my bad. PDFtamplate is not pretty, just proof of concept. Some closing buttons still bugy.

+++ Service/contractorservice.cpp
@@ -1,5 +1,8 @@
 #include "contractorservice.h"
 #include "Repository/contractorrepository.h"
+#include "Service/invoiceservice.h"
+
+#include <Repository/invoicerepository.h>
 
 ContractorService::ContractorService() {}
 
@@ -22,6 +25,9 @@ bool ContractorService::AddContractor(ContractorModel ct)
 
 bool ContractorService::DeleteContractor(ContractorModel ct)
 {
+    if(InvoiceRepository().GetInvoiceByContractor(ct.getId()).size()>0){
+        return false;
+    }
     return ContractorRepository().DeleteContractor(ct);
 
 }
